fix(home): load projects server side for landing wrapper
- fetch projects in the home page route and pass typed data into the landing wrapper
- convert the landing wrapper and projects section to client components that consume props
