+++ src/views/global/home/Projects.tsx
@@ -1,4 +1,6 @@
-﻿// Components
+﻿'use client';
+
+// Components
 import Container from '@/component/Container';
 import LargeLabel from '@/component/LargeLabel';
 import SmallLabel from '@/component/SmallLabel';
@@ -8,31 +10,13 @@ import ProjectCard from '../projects/ProjectCard';
 // Types
 import type { ProjectsType } from '@/types/ProjectsType';
 
-// Actions
-import { getProjects } from '@/app/actions/getProjects';
-
 // Utils
 import { columnise } from '@/utils/columnise';
 
 const columnOffsetClasses = ['md:mt-16', '', 'md:mt-28'];
 
-const Projects = async () => {
-  const rows = await getProjects();
-
-  const projects: ProjectsType[] = rows.map((p) => ({
-    id: p.id,
-    name: p.name,
-    description: p.description,
-    requesterName: p.requesterName ?? undefined,
-    technologies: p.technologiesLabel,
-    year: p.yearLabel,
-    viewCount: p.viewCountLabel,
-    projectLink: p.projectLink ?? undefined,
-    photo: p.photo,
-  }));
-
+const Projects = ({ projects }: { projects: ProjectsType[] }) => {
   const hasData = projects.length > 0;
-
   const columnisedProjects = hasData ? columnise(projects, 3) : [];
 
   return (
